Invert dependency between definitions and notify packages
