+++ definition/compat_test.go
@@ -1,11 +1,9 @@
 package definition
 
 import (
-	"encoding/json"
 	"fmt"
 	"testing"
 
-	"github.com/prometheus/alertmanager/config"
 	"github.com/stretchr/testify/require"
 )
 
@@ -134,36 +132,6 @@ func TestGrafanaToUpstreamConfig(t *testing.T) {
 	for i, r := range cfg.Receivers {
 		require.Equal(t, r.Name, upstream.Receivers[i].Name)
 	}
-}
-
-func TestPostableApiReceiverToApiReceiver(t *testing.T) {
-	postableReceiver := &PostableApiReceiver{
-		Receiver: config.Receiver{
-			Name: "test",
-		},
-		PostableGrafanaReceivers: PostableGrafanaReceivers{
-			GrafanaManagedReceivers: []*PostableGrafanaReceiver{{
-				UID:                   "abc",
-				Name:                  "test",
-				Type:                  "slack",
-				DisableResolveMessage: true,
-				Settings:              RawMessage{'b', 'y', 't', 'e', 's'},
-				SecureSettings:        map[string]string{"key": "value"},
-			}},
-		},
-	}
-	receiver := PostableAPIReceiverToAPIReceiver(postableReceiver)
-
-	require.Equal(t, "test", receiver.Name)
-	require.Equal(t, 1, len(receiver.GrafanaIntegrations.Integrations))
-
-	i := receiver.GrafanaIntegrations.Integrations[0]
-	require.Equal(t, "abc", i.UID)
-	require.Equal(t, "test", i.Name)
-	require.Equal(t, "slack", i.Type)
-	require.Equal(t, true, i.DisableResolveMessage)
-	require.Equal(t, json.RawMessage{'b', 'y', 't', 'e', 's'}, i.Settings)
-	require.Equal(t, map[string]string{"key": "value"}, i.SecureSettings)
 }
 
 const testConfigWithoutGlobal = `
